requirements.txt
streamlit
pillow
pytesseract
gspread
oauth2client
pandas
